+++ crates/cargo-contract/tests/decode.rs
@@ -170,88 +170,3 @@ fn decode_works() {
         .failure()
         .stderr(predicates::str::contains(error_msg));
 }
-
-#[test]
-fn encode_works() {
-    // given
-    let contract = r#"
-		#![cfg_attr(not(feature = "std"), no_std)]
-
-        #[ink::contract]
-        mod incrementer {
-            #[ink(storage)]
-            pub struct Incrementer {
-                value: i32,
-            }
-
-            impl Incrementer {
-                #[ink(constructor)]
-                pub fn new(init_value: i32) -> Self {
-                    Self { value: init_value }
-                }
-
-                #[ink(constructor)]
-                pub fn new_default() -> Self {
-                    Self::new(Default::default())
-                }
-
-                #[ink(message, selector = 0xBABABABA)]
-                pub fn inc(&mut self, by: i32) {
-                    self.value += by;
-                }
-
-                #[ink(message, selector = 0xCACACACA)]
-                pub fn get(&self) -> i32 {
-                    self.value
-                }
-            }
-        }"#;
-
-    let tmp_dir = tempfile::Builder::new()
-        .prefix("cargo-contract.cli.test.")
-        .tempdir()
-        .expect("temporary directory creation failed");
-
-    // cargo contract new decode_test
-    cargo_contract(tmp_dir.path())
-        .arg("new")
-        .arg("incrementer")
-        .assert()
-        .success();
-
-    let project_dir = tmp_dir.path().to_path_buf().join("incrementer");
-
-    let lib = project_dir.join("lib.rs");
-    std::fs::write(lib, contract).expect("Failed to write contract lib.rs");
-
-    tracing::debug!("Building contract in {}", project_dir.to_string_lossy());
-    cargo_contract(&project_dir).arg("build").assert().success();
-
-    // when
-    let output: &str = r#"Encoded data: BABABABA05000000"#;
-
-    // then
-    // message selector and data are being encoded properly
-    cargo_contract(&project_dir)
-        .arg("encode")
-        .arg("--message")
-        .arg("inc")
-        .arg("--args")
-        .arg("5")
-        .assert()
-        .success()
-        .stdout(predicates::str::contains(output));
-
-    // when
-    let output: &str = r#"Encoded data: CACACACA"#;
-
-    // then
-    // message selector is being encoded properly
-    cargo_contract(&project_dir)
-        .arg("encode")
-        .arg("--message")
-        .arg("get")
-        .assert()
-        .success()
-        .stdout(predicates::str::contains(output));
-}
